Add wrq:base_uri/1 and tests
This patch adds a base_uri function to the wrq module. This is
convenient for applications which need to embed a URI in responses.

The new function is also now used by the location header setting code in
webmachine_request. Note that the previous implementation had a bug
that caused the address to be listed in reverse (e.g. com.somewhere)
and default ports to always be included.