Custom Server header support.
The idea behind this extension is that in some cases it is worthwhile to
obscure the actual server name (for example, for an extra layer of
attack protection) or change it any other way.

Of course this is doable with some HTTP proxies such as nginx. However,
in case of nginx (for example) one needs to build it with a custom
third-party module, which complicates the deployment insanely. Besides,
I believe there is a slight performance penalty when rewriting HTTP
headers. So, why not generate them the way you need them right from the
beginning?

This change allows setting custom server name by setting server_name
environment variable for webmachine application. By default, original
server name will be used.