improve new_webmachine.sh support for yaws
Add a rebar.config.script template to priv/templates that changes the
webmachine dependency for apps created by new_webmachine.sh to the
vinoski/webmachine fork if yaws is the selected web server.

Add a new variable webserver_app to priv/templates/wmskel.template to allow
the dependencies in a new webmachine application's .app file to be set
properly for the chosen web server.