conditional compilation for webmachine_yaws module
To use this webmachine fork with mochiweb, the webmachine_yaws module must
compile successfully if no yaws code is available. Use rebar.config.script
to inject a WEBMACHINE_YAWS define when the OS environment variable
WEBMACHINE_SERVER=yaws is set. If this define is not set, webmachine_yaws
compiles to an empty module.

Also modify the Makefile to set the WEBMACHINE_SERVER OS env var when
invoking rebar.