Repair and refactoring of the etag eqc test
Change the property to use "*" as the value for the 'If' header if the
list of values generated for that header is empty. This simultaneously
corrects a bad expectation by the property and expands the scope of
the test. Also refactor the eunit setup to use a test suite and use
the eqc:testing_time function to execute for 10 seconds as oppoosed to
a fixed 250 iterations.