Allow multiple webmachine instance to run in the same app
Changes to allow multiple instances of webmachine to coexist within
the same erlang application. The instances can use distinct dispatch
tables and be configured to listen on different IP addresses or port
numbers. The current behavior is maintained as the default, but maybe
overridden by setting default_router to false in the application
environment and starting named router processes from the application
supervisor prior to calling webmachine_mochiweb:start.