Allow the response behavior of 405 Method Not Allowed errors to be changed
The default behavior for 405 errors is to return an empty body. The
change makes it possible to override that behavior using the
error_response_on_405 application environment variable. If this
variable is set to true then error_response is called and the response
may be defined by the configured error handler module.