Factor out ensure_all_started into webmachine_util
We have a special version of ensure_all_started that uses the
default Erlang implementation if it's available, and otherwise falls
back to our own reimplemented version of the function.

This is useful to have in places other than just the one test function,
so I'm factoring this out into the webmachine_util module.