If resource returns {false, ...} for is_authorized, return 401 error.
The wiki mentions that we should return {true, RD, Ctx} on authorized
calls and {false, RD, Ctx} on calls that are not authorized. Returning
false results in 500 internal server error when wrq module tries to
set response headers. Ensure that we are checking for false along with
true and return 401 error.