Update mochiweb dependency to git
The latest version of mochiweb can't be found in hex and the current
project owner does not have control to publish a new version. In order
to have webmachine be more compatible with erlang 21, the mochiweb version
needs to be upgraded as the "tuple calls" have officially been removed
from the language.