return default peer on empty x-forwarded-for
As reported in #174, an empty X-Forwarded-For header would cause a crash
in `x_peername` at the call to `lists:last/1`. This change checks whether
the header is empty, and returns the `Default` value if so to avoid this
crash.

The check is made after the tokenization, because otherwise a header value
of ",,,,," would also produce an empty list for lists:last to crash on.