prevent access logger from barfing on missing end time
In some error cases, like when a request does not match any dispatch rule,
no `end_time` is set in the requests's log data. If `RequestTiming` was
enabled in `webmachine_access_log_handler`, this would pass `undefined` to
`timer:now_diff/2`, causing the process to exit with reason
`function_clause`.

This patch substitutes the current time for a missing end time to prevent
the exit from happening.