ignore empty multipart preamble
Issue #107 noted that Riak begins its link-walk multipart response with a
CRLF. According to section 7.2.1 of RFC1341, this is correct - it marks
the end of an empty preamble. But it caused a bad-match when `make_part`
(formerly called at line 75, now at line 77) tried to parse it has a part.

This change ignores an empty preamble by treating a bare CRLF just like we
treat the empty binary produces by the `re:split` at the start of the
stream - by throwing it away and moving on.

This is probably still incorrect. It correctly parses the example given in
the bug report. But I have a feeling that the right sequence of `Hunk`s
lining up on a place in a valid body where a `<<"\r\n">>` just happens to
be could fail. A complete fix looks like it will be more invasive, so this
patches the obvious for now.