do not include rebar3 in the repo
The Makefile prefers the rebar3 in this directory to any available on
$PATH. This caused a problem for me when I upgraded to OTP R25, which the
included rebar didn't support.

It seems like it was included primarily to support an older testing
environment (Travis). Now that that's not necessary, remove it so that
devs can maintain their own installation. The download in the Makefile
appears to still work for those that haven't installed rebar3 on their
systems yet.