Include custom status phrase in error body
If a resource uses a `{halt, {Code, Phrase}}` return value, where Code is
>= 400, make sure that Phrase ends up in the body of the response. (It
already ends up in the status line.) See #242.

This required changing the error logger interface to expect `{Code,
Phrase}`. That will break custom error loggers, but there is no access to
to the custom phrase otherwise - `wrq` stores it, but strips it off in the
accessor function.