replace webmachine_error with webmachine_status_code
The `webmachine_error` module was used to fill in the body of an error
response in cases where `webmachine_error_handler` didn't have a
predefined message to render. It was mostly duplication of
`httpd_util:reason_phrase/1`, though. It had been proposed as a place to
store additional status code phrases, but now that custom phrases make it
all the way to the error handler rendering function, it seems unnecessary.

One case where `webmachine_error` improved on `http_util:reason_phrase/1`
is in the phrase produced for extension 4xx codes. The latter produces
"Internal Server Error", but the former produced "Client Error". The
`webmachine_status_code` module is introduced in this change to retain and
expand this improvement. Instead of specifically producing "Client Error"
for 4xx codes, it instead produces the x00 code of the same class for any
unknown status code. That is, where `webmachine_error` would have produced
"418 Client Error", `webmachine_status_code` produces "418 Bad Request".
This seems inline with RFC 2616's section 6.1.1 statement that,
"applications MUST treat any unrecognized response as being equivalent to
the x00 status code of that class