drop support for OTP 18/19
Github is deprecating actions support for Ubuntu 18.04. See the build/test
run on #234 and https://github.com/actions/runner-images/issues/6002

OTP 18 and 19 have been the cause of a few fixup commits in recent changes
anyway, because of missing functions. Time to get rid of them.

This also removes a long unneeded bit of rebar config related to R15.