fix error handler body content test (#326)
While piping custom status phrases through to the error handler in #323,
an extra assertion was added to a test that was going to receive an error
status that `webmachine_error_handler` would render a message body for.
Unfortunately, that test communicated with Webmachine via a `gen_tcp`
socket directly, instead of an http client. So occasionally, it would read
only part of the response, because that's all that would show up in the
first packet.

This change makes the test for rendered error body its own test, and uses
httpc to make sure that the body is actually read before trying to match
it.