throw an exception when a socket error is encountered on recv (#248)
If a client disconnects before finishing sending the request body, an
error may be returned from `mochiweb_socket:recv`. That used to raise a
`badmatch` exception here. A resource may prefer to learn about the
disconnection. This change introduces a thrown `webmachine_recv_error`
exception, which is a more stable API than trying to catch a `badmatch`.
If a resource doesn't catch the throw, then it still tears down as
expected.