nearly eliminate stack trace for socket receive errors
The previous commit introduced an easy to detect thrown exception when a
socket error is encountered while reading the request body. This commit
demonstrates how that exception can be caught, and reduces the stack trace
in the logged error to just the first location in the resource function
that was called. These socket errors are usually not bugs in the resource,
but instead are clients behaving strangely, so logging lots of details
about the resource may not make sense.