ensure request body is read
Both #168 and #232 are caused by mochiweb closing the socket after
webmachine finishes processing a request without reading the body. This
change makes sure that the body has been read, except for failures that
happen before resource init (like dispatch).

To support finishing a stream, if the resource stops in the middle, the
"Next" fun is stashed in the process dictionary. An alternate fix would
have been to stash the number of bytes left in the stream/chunk, and start
a fresh "Next" chain, but that felt somehow dirtier.

If the request cannot be read, "Connection: close" is added to the
response headers. This is kind of advisory, but an attempt to help clients
expect the socket close headed their way.