test request body consumption via multiple request tests
This change introduces a new suite of tests that deliver two requests at
once on the same connection. The test resource that receives the request
either, depending on the path and method, either reads the body, reads
part of the body, or does not read any of the body. This is to exercise
the new `maybe_flush_request_body` feature.

In all but one test, the expectation is that webmachine flushes the
request body and handles the second request correctly. Before the flush
was introduced, these four tests would fail because the unconsumed body in
the first request would cause mochiweb to close the connection instead of
moving on to the second request:

  * `unauthorized_post_then_get`
  * `too_large_unauthorized_post_then_get`
  * `half_chunk_then_get`
  * `half_unchunk_then_get`

The `too_large_unauthorized_post_then_get` test technically fails before
the introduction of `maybe_flush_request_body` because the test can't read
the non-existent configuration variable it tries to save and restore. It's
also the only test that does not expect the second request to be handled.
It tests that a client can't cause webmachine to hang up flushing an
infinitely long unconsumed request body.

The `delete_with_body_then_get` test did not fail before the introduction
of `maybe_flush_request_body` because consuming a body in DELETE was
handled specially. I didn't find a test that confirmed that, so this one
is introduced.

The remaining two tests are "happy path" tests, where there is no
unconsumed request body left to flush after handling the first request.
These should have always worked in webmachine, but I didn't find a test
verifying that.

This change also removes the "same_inets..." test introduced with the
maybe-flush fix. That test was a direct translation of the reported issue
that led to the fix, but it was flaky. These tests cover the same ground
without the flakiness.