add expect-100 testing to multi-request test
As noted in #168, flushing the request body when expect-100-continue is in
play is tricky. These two tests verify that the new flush code works. Both
tests send their first request with "Expect: 100-continue". One test hits
a resource path that does not read the body, and verifies that no "100
Continue" arrives. The second test hits a resource path that does not
finish reading the body, and verifies that "100 Continue" does arrive.
Both tests expect the second request on the connection to be handled
correctly.