stop and close connection on body-sending error
If an error occurs while sending the body of the response, just stop and
close the connection. This prevents the problem reported in issue #60,
where an error in a streaming response body causes the stream to be
evaluated twice (the second time came after the error handler was engaged,
and returned the same stream-body definition, because it couldn't tell
whether the resource rendered its own error message).

This is a little more invasive than just catching any errors from
`send_body`, in order to report the number of bytes actually sent more
accurately in the access logger.

While an error message is no longer included in the response body, an
error is logged by the error log handler.