do not call log_error from error_handler; use notes
This is two changes in one:
  1. Replace the `log_error` event form that includes Code/Req/Reason with
an `error` note in the request data. This arrives at the error logger in
the `notes` field of `#wm_log_data` included with the `log_access` event.
  2. Remove the call to `webmachine_log` from `webmachine_error_handler`.
The `webmachine_error_handler` is now purely an error message renderer,
for response bodies. Adding the note to the log data is the responsibility
of the code that caught the error.